+++ services/liaradb/collection/keyvalue/key_value.go
@@ -71,8 +71,7 @@ func (kv *KeyValue) getItem(ctx context.Context, tn tablename.TableName, rid lin
 		return nil, page.ErrNotPage
 	}
 
-	// TODO: Fix this type
-	d, ok := n.Child(int16(rid.Position()))
+	d, ok := n.Child(rid.Position())
 	if !ok {
 		return nil, btree.ErrNotFound
 	}
