+++ services/liaradb/collection/outbox/outbox.go
@@ -79,8 +79,7 @@ func (o *Outbox) getItem(ctx context.Context, tn tablename.TableName, rid link.R
 		return nil, page.ErrNotPage
 	}
 
-	// TODO: Fix this type
-	d, ok := n.Child(int16(rid.Position()))
+	d, ok := n.Child(rid.Position())
 	if !ok {
 		return nil, btree.ErrNotFound
 	}
